fix fileParser.h

+++ include/fileParser.h
@@ -8,23 +8,24 @@
 typedef struct __attribute__((packed))
 {
     char type[2];   /* Magic identifier            */
-    size_t size; /* File size in bytes          */
-    uint16_t reserved1, reserved2;
-    uint32_t offset; /* Offset to image data, bytes */
+    u_int32_t size; /* File size in bytes          */
+    u_int16_t reserved1, reserved2;
+    u_int32_t offset; /* Offset to image data, bytes */
 } HEADER;
 
 typedef struct __attribute__((packed))
 {
-    size_t size;                       /* Header size in bytes      */
-    int32_t width, height;             /* Width and height of image */
-    uint16_t planes;                   /* Number of colour planes   */
-    uint16_t bits;                     /* Bits per pixel            */
-    uint32_t compression;              /* Compression type          */
-    size_t imageSize;                  /* Image size in bytes       */
-    uint32_t xResolution, yResolution; /* Pixels per meter          */
-    uint32_t nColors;                  /* Number of colours         */
-    uint32_t importantColors;          /* Important colours         */
+    u_int32_t size;                     /* Header size in bytes      */
+    int32_t width, height;              /* Width and height of image */
+    u_int16_t planes;                   /* Number of colour planes   */
+    u_int16_t bits;                     /* Bits per pixel            */
+    u_int32_t compression;              /* Compression type          */
+    u_int32_t imageSize;                /* Image size in bytes       */
+    u_int32_t xResolution, yResolution; /* Pixels per meter          */
+    u_int32_t nColors;                  /* Number of colours         */
+    u_int32_t importantColors;          /* Important colours         */
 } INFO_HEADER;
+
 
 typedef struct
 {
